Remove outer for loop in GeneratePossibleNotes
Due to changing the iterator, the outer for loop is not needed. The
inner for loop checks the depth of the tree_iter to determine if the
end of the cantus firmus has been reached. Notes added to the tree are
based on cantus_firmus_[current_depth], not cantus_firmus_[i].